Rename routes reducer to nav
