+++ source/reducers/index.js
@@ -1,7 +1,7 @@
 import { combineReducers } from "redux";
 import archives from "./archives.js";
 import archiveContents from "./archiveContents.js";
-import nav from "./routes.js";
+import nav from "./nav.js";
 import addArchive from "./addArchive.js";
 import remoteExplorer from "./remoteExplorer.js";
 import entry from "./entry.js";
